requirements.txt
# ==================== CORE ====================
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
pydantic-settings>=2.0.0

# ==================== SEGURIDAD ====================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ==================== BASE DE DATOS ====================
alembic>=1.13.0

# ==================== TESTING ====================
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0

# ==================== FORMATO Y CALIDAD ====================
black>=24.0.0
isort>=5.0.0